+++ src/main/java/com/racingcar/model/RacingCar.java
@@ -1,7 +1,7 @@
 package com.racingcar.model;
 
-import com.racingcar.service.DefaultRacingCarRule;
 import com.racingcar.service.RacingCarMoveStrategy;
+import com.racingcar.service.RacingCarRule;
 
 public class RacingCar {
 
@@ -11,13 +11,17 @@ public class RacingCar {
 
     public RacingCar(Integer racingStatus) {
         this.racingStatus = racingStatus;
-        this.racingCarMoveStrategy = new DefaultRacingCarRule();
+        this.racingCarMoveStrategy = new RacingCarRule();
     }
 
     public RacingCar(Integer racingStatus, String racingCarName) {
         this.racingStatus = racingStatus;
         this.racingCarName = racingCarName;
-        this.racingCarMoveStrategy = new DefaultRacingCarRule();
+        this.racingCarMoveStrategy = new RacingCarRule();
+    }
+
+    public void setRacingStatus(Integer racingStatus) {
+        this.racingStatus = racingStatus;
     }
 
     public Integer getRacingStatus() {
